Add files via upload
Yo you can add links for the the other things on the page. Next week we can review the website . If you host it online too that would be great. Feel free to change up the pics on the gallery on the frontage to our groups actual corresponding chair topics and add a bit more to the description